+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Release Notes
 All notable changes to this project will be documented in this file.
 
+## 0.26.1 – May 4, 2020
+### Fixed
+- Aggregates on nested array fields failed with `value.substr is not function`.
+- Slow detector for `MIN` `MAX` aggregates must use a specified field as `order by` to detect fast query. 
+- Indexes reloaded from db first time on demand and then every 1 hour. 
+- Config p17 stake types.
+
 ## 0.26.0 – May 2, 2020
 ### New
 - companion fields `*_string` for fields that holds unix time values